# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Karpos API
  version: 0.1.0
  description: REST API de Karpos — plataforma SaaS para manejo integral de cultivos frutales.
  contact:
    email: api@karpos.example
  license:
    name: Karpos Commercial License
servers:
  - url: https://api.karpos.example/v1
    description: Production
  - url: https://api.staging.karpos.example/v1
    description: Staging
  - url: http://localhost:4000/v1
    description: Local development
security:
  - bearerAuth: []
tags:
  - name: predios
  - name: bitacora-verde
  - name: fenoflow
  - name: sanidad-plus
  - name: cosecha360
  - name: karpos-iq
  - name: billing

paths:
  /farms:
    get:
      tags: [predios]
      summary: Lista fincas del tenant
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FarmPage'
    post:
      tags: [predios]
      summary: Crea una finca
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFarm'
      responses:
        '201':
          description: Creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Farm'

  /farms/{id}:
    get:
      tags: [predios]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Farm'

  /plots:
    get:
      tags: [predios]
      parameters:
        - { in: query, name: farmId, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Plot' }

  /field-operations:
    get:
      tags: [bitacora-verde]
      parameters:
        - { in: query, name: plotId, schema: { type: string, format: uuid } }
      responses: { '200': { description: OK } }
    post:
      tags: [bitacora-verde]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogOperation'
      responses: { '201': { description: Created } }

  /phenology/events:
    get:
      tags: [fenoflow]
      parameters:
        - { in: query, name: plotId, required: true, schema: { type: string, format: uuid } }
      responses: { '200': { description: OK } }
    post:
      tags: [fenoflow]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordPhenologyEvent'
      responses: { '201': { description: Created } }

  /phenology/gdd:
    get:
      tags: [fenoflow]
      parameters:
        - { in: query, name: plotId, required: true, schema: { type: string, format: uuid } }
        - { in: query, name: from, required: true, schema: { type: string, format: date } }
        - { in: query, name: to, required: true, schema: { type: string, format: date } }
        - { in: query, name: baseTempC, schema: { type: number, default: 10 } }
        - { in: query, name: upperTempC, schema: { type: number } }
      responses: { '200': { description: OK } }

  /health/sprays:
    post:
      tags: [sanidad-plus]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordSpray'
      responses: { '201': { description: Created } }

  /health/sprays/{id}/void:
    post:
      tags: [sanidad-plus]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties: { reason: { type: string, minLength: 5 } }
      responses: { '200': { description: OK } }

  /harvest/lots:
    post:
      tags: [cosecha360]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordHarvestLot'
      responses: { '201': { description: Created } }

  /copilot/ask:
    post:
      tags: [karpos-iq]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question]
              properties:
                sessionId: { type: string, format: uuid }
                question: { type: string, minLength: 2, maxLength: 4000 }
                context:
                  type: object
                  properties:
                    farmId: { type: string, format: uuid }
                    plotId: { type: string, format: uuid }
                    species: { type: string }
      responses: { '200': { description: OK } }

  /billing/webhooks/stripe:
    post:
      tags: [billing]
      security: []
      summary: Stripe webhook
      requestBody:
        required: true
        content: { application/json: { schema: { type: object } } }
      responses: { '200': { description: OK } }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Farm:
      type: object
      properties:
        id: { type: string, format: uuid }
        orgId: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        countryCode: { type: string, minLength: 2, maxLength: 2 }
        region: { type: string, nullable: true }
        timezone: { type: string }
    CreateFarm:
      type: object
      required: [code, name, countryCode, timezone]
      properties:
        code: { type: string }
        name: { type: string }
        countryCode: { type: string }
        region: { type: string }
        locality: { type: string }
        timezone: { type: string }
        elevationM: { type: integer }
        totalAreaHa: { type: number }
    FarmPage:
      type: object
      properties:
        rows: { type: array, items: { $ref: '#/components/schemas/Farm' } }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
    Plot:
      type: object
      properties:
        id: { type: string, format: uuid }
        farmId: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        speciesId: { type: string, format: uuid }
        status: { type: string, enum: [planning, establishing, active, retired] }
    LogOperation:
      type: object
      required: [plotId, operationTypeId, performedAt]
      properties:
        plotId: { type: string, format: uuid }
        operationTypeId: { type: string, format: uuid }
        performedAt: { type: string, format: date-time }
        crewId: { type: string, format: uuid }
        workersCount: { type: integer }
        areaCoveredHa: { type: number }
    RecordPhenologyEvent:
      type: object
      required: [plotId, observedAt, bbchCode]
      properties:
        plotId: { type: string, format: uuid }
        observedAt: { type: string, format: date-time }
        bbchCode: { type: string, pattern: '^\d{2}$' }
        observedPct: { type: number }
        notes: { type: string }
    RecordSpray:
      type: object
      required: [plotId, appliedAt, areaTreatedHa, items]
      properties:
        plotId: { type: string, format: uuid }
        appliedAt: { type: string, format: date-time }
        finishedAt: { type: string, format: date-time }
        areaTreatedHa: { type: number }
        applicatorName: { type: string }
        applicatorLicense: { type: string }
        equipment: { type: string }
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [productId, doseValue, doseUnit]
            properties:
              productId: { type: string, format: uuid }
              targetTaxonId: { type: string, format: uuid }
              doseValue: { type: number }
              doseUnit: { type: string }
    RecordHarvestLot:
      type: object
      required: [plotId, lotCode, harvestedOn, grossKg]
      properties:
        plotId: { type: string, format: uuid }
        lotCode: { type: string }
        harvestedOn: { type: string, format: date }
        grossKg: { type: number }
        tareKg: { type: number }
        containersCount: { type: integer }